make manual payments work again
When I added the JS alert to choose the amount you paid when you click the "OK, I paid" button, it turns out it broke manual payments on the admin page -- this is because both inputs have the "new_payment" id, despite being on different pages and having slightly different functions. Got around this by specifying the new_payment button on the home page in the JQuery selector.